Adição de tela principal e correção de alguns erros
