+++ codigo/Prototipo/Cadastro_Med.html
@@ -6,86 +6,84 @@
     <link href="assets/css/main.css" rel="stylesheet">
     <link href="assets/css/bootstrap.min.css" rel="stylesheet">
     <title>Santé</title>
-    <link rel="icon" href="http://127.0.0.1:5500/assets/img/Saude2.png">
+    <link rel="icon" href="assets/img/Saude2.png">
 </head>
 <body>
-    
-    <Header>
+    <header>
         <div class="header">
-          <a href="index.html" class="logo">Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
-          <div class="header-right">
-            <a href="#home">Meu Caso</a>
-            <a href="#contact">Quem Somos</a>
-            <button href="#" type="button" class="btn-azul">Cadastrar</button>
-            <button href="#" type="button" class="btn-azul">Login</button>
-          </div>
+            <a href="index.html" class="logo">Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
+            <div class="header-right">
+                <a href="#home">Meu Caso</a>
+                <a href="#contact">Quem Somos</a>
+                <button type="button" class="btn-azul">Cadastrar</button>
+                <button type="button" class="btn-azul">Login</button>
+            </div>
         </div>
-    </Header>
+    </header>
 
-    <br><div class="container shadow p-3 mb-5 bg-body-tertiary rounded">
+    <br>
+    <div class="container shadow p-3 mb-5 bg-body-tertiary rounded">
         <form>
-          <h2 class="d-flex justify-content-center">Cadastro de Médico</h2>
-          <div class="form-group">
-            <label for="nome">Nome:</label>
-            <input type="text" class="form-control" id="nome" name="nome" placeholder="Matheus" required>
-          </div>
-          <div class="form-group">
-            <label for="crm">CRM:</label>
-            <input type="text" class="form-control" id="crm" name="crm" placeholder="000000/SP" required>
-          </div>
-          <div class="form-group">
-            <label for="especialidade">Especialidade:</label>
-            <select class="form-control" id="especialidade" name="especialidade" placeholder="Cardiologista" required>
-              
-            </select>
-            
-          </div>
-          <div class="form-group">
-            <label for="telefone">Telefone:</label>
-            <input type="tel" class="form-control" id="telefone" name="telefone" placeholder="(00) 0 0000-000" required>
-          </div>
-          <div class="form-group">
-            <label for="email">Email:</label>
-            <input type="email" class="form-control" id="email" name="email" placeholder="matheus@gmail.com"required>
-          </div>
-          <div class="form-group">
-            <label for="senha">Senha:</label>
-            <input type="password" class="form-control" id="senha" name="Senha" placeholder="******" required>
-          </div>
-          <div class="form-group">
-            <label for="data_nascimento">Data de Nascimento:</label>
-            <input type="date" class="form-control" id="data_nascimento" name="data_nascimento" required>
-          </div><br>
-          <div class="d-flex justify-content-center">
-            <a href="#" onclick="armazenaDados()"><button type="submit" class="btn btn-primary">Cadastrar</button></a>
-            <button onclick="window.location.href='index.html'" class="btn btn-info">Voltar</button>
-          </div>
-          <br><a href="Login_Med.html" class="d-flex justify-content-center link" >Já tenho uma conta</a>
+            <h2 class="d-flex justify-content-center">Cadastro de Médico</h2>
+            <div class="form-group">
+                <label for="nome">Nome:</label>
+                <input type="text" class="form-control" id="nome" name="nome" placeholder="Matheus" required>
+            </div>
+            <div class="form-group">
+                <label for="crm">CRM:</label>
+                <input type="text" class="form-control" id="crm" name="crm" placeholder="000000/SP" required>
+            </div>
+            <div class="form-group">
+                <label for="especialidade">Especialidade:</label>
+                <select class="form-control" id="especialidade" name="especialidade" required>
+                </select>
+            </div>
+            <div class="form-group">
+                <label for="telefone">Telefone:</label>
+                <input type="tel" class="form-control" id="telefone" name="telefone" placeholder="(00) 0 0000-000" required>
+            </div>
+            <div class="form-group">
+                <label for="email">Email:</label>
+                <input type="email" class="form-control" id="email" name="email" placeholder="matheus@gmail.com" required>
+            </div>
+            <div class="form-group">
+                <label for="senha">Senha:</label>
+                <input type="password" class="form-control" id="senha" name="senha" placeholder="******" required>
+            </div>
+            <div class="form-group">
+                <label for="data_nascimento">Data de Nascimento:</label>
+                <input type="date" class="form-control" id="data_nascimento" name="data_nascimento" required>
+            </div>
+            <br>
+            <div class="d-flex justify-content-center">
+                <button type="submit" onclick="armazenaDados()" class="btn btn-primary">Cadastrar</button>
+                <button type="button" onclick="window.location.href='index.html'" class="btn btn-info">Voltar</button>
+            </div>
+            <br>
+            <a href="Login_Med.html" class="d-flex justify-content-center link">Já tenho uma conta</a>
         </form>
-      </div>
-
-  <!--Footer das informações do site -->
-  <footer>
-    <div class="footer">
-      <div class="row g-0 text-center">
-
-        <div class="col-6 col-md-4 p-2">
-          <h5>Seja Nosso cliente</h5>
-          <a href='Cadastro_Med.html' style="color: white;">Cadastre-se</a>
-        </div>
-        <div class="col-6 col-md-4 p-2">
-          <h5>Fale com a gente</h5>
-          <p>Enviar Mensagem</p>
-          <p>Suporte (31) 9 9999-9999</p>
-          <p>Reclame aqui</p>
-        </div>
-        <div class="col-6 col-md-4 p-2">
-          <h5>Criadores</h5>
-          <a href="https://github.com/ICEI-PUC-Minas-PMGES-TI/pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar"style="color: white;">Equipe de Diw G7 - Puc Minas Coração Eucaristico </a>
-        </div>
-      </div>
     </div>
-  </footer>
+
+    <footer>
+        <div class="footer">
+            <div class="row g-0 text-center">
+                <div class="col-6 col-md-4 p-2">
+                    <h5>Seja Nosso cliente</h5>
+                    <a href="Cadastro_Med.html" style="color: white;">Cadastre-se</a>
+                </div>
+                <div class="col-6 col-md-4 p-2">
+                    <h5>Fale com a gente</h5>
+                    <p>Enviar Mensagem</p>
+                    <p>Suporte (31) 9 9999-9999</p>
+                    <p>Reclame aqui</p>
+                </div>
+                <div class="col-6 col-md-4 p-2">
+                    <h5>Criadores</h5>
+                    <a href="https://github.com/ICEI-PUC-Minas-PMGES-TI/pmg-es-2024-1-ti1-2401100-g7-saude-e-bem-estar" style="color: white;">Equipe de Diw G7 - Puc Minas Coração Eucaristico</a>
+                </div>
+            </div>
+        </div>
+    </footer>
     <script src="assets/js/santelib.js"></script>
     <script src="assets/js/main.js"></script>
 </body>
