+++ codigo/Prototipo/Calendario_sin.html
@@ -17,12 +17,11 @@
             <div class="header-right">
                 <a href="#home">Meu Caso</a>
                 <a href="#contact">Quem Somos</a>
-
                 <div class="dropdown header-right">
                     <a class="dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-bs-toggle="dropdown" aria-expanded="false">
                       <img src="assets/img/perfil.png" alt="saude" class="img-saude">
                     </a> 
-                    <ul class="dropdown-menu" aria-labelledby="dropdownMenuLink" >
+                    <ul class="dropdown-menu" aria-labelledby="dropdownMenuLink">
                       <li><a class="dropdown-item" href="#">Configuração</a></li>
                       <li><a class="dropdown-item" href="index.html">Sair</a></li>
                     </ul>
