+++ codigo/Prototipo/Perfil_Med.html
@@ -6,56 +6,63 @@
     <link href="assets/css/main.css" rel="stylesheet">
     <link href="assets/css/perfil_med.css" rel="stylesheet">
     <link href="assets/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="icon" href="http://127.0.0.1:5500/assets/img/Saude2.png">
+    <link rel="icon" href="assets/img/Saude2.png">
     <title>Santé</title>
 </head>
 <body>
     <header>
         <div class="header">
-            <a href="index.html" class="logo"> Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
+            <a href="index.html" class="logo">Santé <img src="assets/img/Saude2.png" alt="saude" class="img-saude"></a>
             <div class="header-right">
                 <a href="#home">Meu Caso</a>
                 <a href="#contact">Quem Somos</a>
-
                 <div class="dropdown header-right">
                     <a class="dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-bs-toggle="dropdown" aria-expanded="false">
                       <img src="assets/img/perfil.png" alt="saude" class="img-saude">
                     </a> 
                     <ul class="dropdown-menu" aria-labelledby="dropdownMenuLink">
-                      <li><a class="dropdown-item" href="#" >Configurações</a></li>
-                      <li><a class="dropdown-item" id="sair"href="index.html" >Sair</a></li>
+                        <li><a class="dropdown-item" href="#">Configurações</a></li>
+                        <li><a class="dropdown-item" id="sair" href="index.html">Sair</a></li>
                     </ul>
                 </div>
             </div>
         </div>
     </header>
 
-    <main><br>
-            <div class="fundo">
-                <div class="">
-                    <img src="assets/img/perfil.png" alt="perfil_grand" align="left" class="img_perfil"><br>
-                    <h1 id="name_perfil">Medico Name</h1><br>
-                    <h4 id="idade_perfil"></h4>
-                    <h4 id="email_perfil"></h4>
-                    <h4 id="telefone_perfil"></h4>
-                    <h4 id="expecializacao_perfil"></h4><br><br>
-                    <button type="submit" class="btn btn-primary">Editar</button>
-                    <button type="submit" class="btn btn-danger">Excluir dados</button><br><br><br>
-                </div>
+    <main>
+        <br>
+        <div class="fundo">
+            <div>
+                <img src="assets/img/perfil.png" alt="perfil_grand" align="left" class="img_perfil">
+                <br>
+                <h1 id="name_perfil">Medico Name</h1>
+                <br>
+                <h4 id="idade_perfil"></h4>
+                <h4 id="email_perfil"></h4>
+                <h4 id="telefone_perfil"></h4>
+                <h4 id="especializacao_perfil"></h4>
+                <br><br>
+                <button type="button" class="btn btn-primary">Editar</button>
+                <button type="button" class="btn btn-danger">Excluir dados</button>
+                <br><br><br>
             </div>
-        <div class=""><br><br>
+        </div>
+        <div>
+            <br><br>
             <div class="fundo">
-              <div style="margin-left:20px;">
-                <h1 >Sobre Mim</h1><br><br>
-                <button type="submit" class="btn btn-primary btn-left">Salvar</button>
-                <button type="submit" class="btn btn-danger btn-left">Cancelar</button><br><br>
-              </div>
+                <div style="margin-left:20px;">
+                    <h1>Sobre Mim</h1>
+                    <br><br>
+                    <button type="button" class="btn btn-primary btn-left">Salvar</button>
+                    <button type="button" class="btn btn-danger btn-left">Cancelar</button>
+                    <br><br>
+                </div>
             </div>
         </div>
     </main>
 
     <footer>
-        <div class="footer ">
+        <div class="footer">
             <div class="row g-0 text-center">
                 <div class="col-6 col-md-4 p-2">
                     <h5>Seja Nosso cliente</h5>
@@ -69,7 +76,7 @@
                 </div>
                 <div class="col-6 col-md-4 p-2">
                     <h5>Criadores</h5>
-                    <p>Equipe de Diw G7 - Puc Minas Coração Eucaristico </p>
+                    <p>Equipe de Diw G7 - Puc Minas Coração Eucaristico</p>
                 </div>
             </div>
         </div>
@@ -78,6 +85,5 @@
     <script src="assets/js/main.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/5.3.0/js/bootstrap.bundle.min.js"></script>
-    
 </body>
 </html>
